CRUD funcionando para Produto e Loja

+++ reobra/src/main/java/br/com/reobra/repository/LojaRepository.java
@@ -1,8 +1,8 @@
 package br.com.reobra.repository;
 
-import br.com.reobra.model.Ocorrencia;
+import br.com.reobra.model.Loja;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LojaRepository extends JpaRepository<Loja, Integer> {
+public interface LojaRepository extends JpaRepository<Loja, Long> {
 
 }
