+++ reobra/src/main/java/br/com/reobra/repository/ProdutoRepository.java
@@ -1,8 +1,8 @@
 package br.com.reobra.repository;
 
-import br.com.reobra.model.AtividadeAgendada;
+import br.com.reobra.model.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
+public interface ProdutoRepository extends JpaRepository<Produto, Long> {
 
 }
